+++ src/constants/threats.js
@@ -1,0 +1,102 @@
+export const threats = [
+  {
+    title: 'Custom',
+    range: '3',
+  },
+  {
+    title: 'SA-2B/F',
+    range: '18.4',
+  },
+  {
+    title: 'SA-2D/E',
+    range: '23.2',
+  },
+  {
+    title: 'SA-3',
+    range: '13',
+  },
+  {
+    title: 'SA-5',
+    range: '162',
+  },
+  {
+    title: 'SA-6',
+    range: '13.4',
+  },
+  {
+    title: 'SA-8',
+    range: '5.5',
+  },
+  {
+    title: 'SA-9',
+    range: '2.3',
+  },
+  {
+    title: 'SA-10A/B',
+    range: '40',
+  },
+  {
+    title: 'SA-11',
+    range: '17',
+  },
+  {
+    title: 'SA-12A',
+    range: '50',
+  },
+  {
+    title: 'SA-12B',
+    range: '80',
+  },
+  {
+    title: 'SA-13',
+    range: '2.7',
+  },
+  {
+    title: 'SA-15',
+    range: '6.5',
+  },
+  {
+    title: 'SA-17',
+    range: '17',
+  },
+  {
+    title: 'SA-19',
+    range: '6.5',
+  },
+  {
+    title: 'SA-20',
+    range: '80',
+  },
+  {
+    title: 'SA-21',
+    range: '64.8',
+  },
+  {
+    title: 'SA-22',
+    range: '10.8',
+  },
+  {
+    title: 'CROTALE/FM-80',
+    range: '15',
+  },
+  {
+    title: 'I-HAWK',
+    range: '21.6',
+  },
+  {
+    title: 'PATRIOT',
+    range: '86.4',
+  },
+  {
+    title: 'RAPIER',
+    range: '3.8',
+  },
+  {
+    title: 'ROLAND-II',
+    range: '4.3',
+  },
+]
+
+export const units = ['NM', 'm', 'km']
+
+export const sovereignties = ['Hostile', 'Suspect', 'Unknown', 'Friendly']